chore(eth): split consensus and evm impls
