+++ lib/ethereum-sync-protocol/Cargo.toml
@@ -13,21 +13,14 @@ test-include = ["lib/ethereum-sync-protocol/src/test/"]
 beacon-api-types            = { workspace = true, features = ["ssz"] }
 bytes                       = { workspace = true }
 ethereum-light-client-types = { workspace = true }
-hash-db                     = { version = "0.16.0", default-features = false }
-hash256-std-hasher          = { version = "0.15.2", default-features = false }
 hex                         = { workspace = true, features = ["alloc"] }
 lazy_static                 = { workspace = true }
-memory-db                   = { version = "0.32.0", default-features = false }
 milagro_bls                 = { workspace = true }
-primitive-types             = { workspace = true, features = ["rlp"] }
-rlp                         = { workspace = true, features = ["derive", "std"] }
 serde                       = { workspace = true, features = ["derive"] }
 serde-utils                 = { workspace = true }
 sha2                        = { workspace = true }
-sha3                        = { workspace = true }
 ssz                         = { workspace = true }
 thiserror                   = { workspace = true }
-trie-db                     = { version = "0.28", default-features = false }
 typenum                     = { workspace = true, features = ["const-generics", "no_std"] }
 unionlabs                   = { workspace = true, features = ["cosmwasm"] }
