+++ lib/ethereum-sync-protocol/src/lib.rs
@@ -2,8 +2,6 @@ extern crate alloc;
 
 pub mod crypto;
 pub mod error;
-pub mod primitives;
-mod rlp_node_codec;
 // REVIEW: Unused?
 pub mod utils;
 
@@ -16,27 +14,23 @@ use beacon_api_types::{
     ChainSpec, DomainType, ExecutionPayloadHeaderSsz, ForkParameters, LightClientHeader,
     SyncCommittee, SyncCommitteeSsz,
 };
-use hash_db::HashDB;
-use memory_db::{HashKey, MemoryDB};
 use ssz::Ssz;
-use trie_db::{Trie, TrieDBBuilder};
 use typenum::Unsigned;
 use unionlabs::{
     bls::{BlsPublicKey, BlsSignature},
     ensure,
-    hash::{H160, H256},
-    uint::U256,
+    hash::H256,
 };
 
 use crate::{
     error::Error,
-    primitives::{Account, GENESIS_SLOT},
-    rlp_node_codec::{keccak_256, EthLayout, KeccakHasher},
     utils::{
         compute_domain, compute_epoch_at_slot, compute_fork_version, compute_signing_root,
         compute_sync_committee_period_at_slot, validate_merkle_branch,
     },
 };
+
+pub const GENESIS_SLOT: u64 = 0;
 
 pub trait BlsVerify {
     fn fast_aggregate_verify<'pk>(
@@ -252,94 +246,6 @@ pub fn validate_light_client_update<C: ChainSpec, V: BlsVerify>(
     )?;
 
     Ok(())
-}
-
-fn get_node(
-    root: H256,
-    key: impl AsRef<[u8]>,
-    proof: impl IntoIterator<Item = impl AsRef<[u8]>>,
-) -> Result<Option<Vec<u8>>, Error> {
-    let mut db = MemoryDB::<KeccakHasher, HashKey<_>, Vec<u8>>::default();
-    proof.into_iter().for_each(|n| {
-        db.insert(hash_db::EMPTY_PREFIX, n.as_ref());
-    });
-
-    let root: primitive_types::H256 = root.into();
-    let trie = TrieDBBuilder::<EthLayout>::new(&db, &root).build();
-    Ok(trie.get(&keccak_256(key.as_ref()))?)
-}
-
-/// Verifies if the `storage_root` of a contract can be verified against the state `root`.
-///
-/// * `root`: Light client update's (attested/finalized) execution block's state root.
-/// * `address`: Address of the contract.
-/// * `proof`: Proof of storage.
-/// * `storage_root`: Storage root of the contract.
-///
-/// NOTE: You must not trust the `root` unless you verified it by calling [`validate_light_client_update`].
-pub fn verify_account_storage_root(
-    root: H256,
-    address: &H160,
-    proof: impl IntoIterator<Item = impl AsRef<[u8]>>,
-    storage_root: &H256,
-) -> Result<(), Error> {
-    match get_node(root, address.as_ref(), proof)? {
-        Some(account) => {
-            let account = rlp::decode::<Account>(account.as_ref()).map_err(Error::RlpDecode)?;
-            ensure(
-                &account.storage_root == storage_root,
-                Error::ValueMismatch {
-                    expected: storage_root.as_ref().into(),
-                    actual: account.storage_root.into(),
-                },
-            )?;
-            Ok(())
-        }
-        None => Err(Error::ValueMissing {
-            value: address.as_ref().into(),
-        })?,
-    }
-}
-
-/// Verifies against `root`, if the `expected_value` is stored at `key` by using `proof`.
-///
-/// * `root`: Storage root of a contract.
-/// * `key`: Padded slot number that the `expected_value` should be stored at.
-/// * `expected_value`: Expected stored value.
-/// * `proof`: Proof that is generated to prove the storage.
-///
-/// NOTE: You must not trust the `root` unless you verified it by calling [`verify_account_storage_root`].
-pub fn verify_storage_proof(
-    root: H256,
-    key: U256,
-    expected_value: &[u8],
-    proof: impl IntoIterator<Item = impl AsRef<[u8]>>,
-) -> Result<(), Error> {
-    match get_node(root, key.to_be_bytes(), proof)? {
-        Some(value) if value == expected_value => Ok(()),
-        Some(value) => Err(Error::ValueMismatch {
-            expected: expected_value.into(),
-            actual: value,
-        })?,
-        None => Err(Error::ValueMissing {
-            value: expected_value.into(),
-        })?,
-    }
-}
-
-/// Verifies against `root`, that no value is stored at `key` by using `proof`.
-///
-/// * `root`: Storage root of a contract.
-/// * `key`: Padded slot number that the `expected_value` should be stored at.
-/// * `proof`: Proof that is generated to prove the storage.
-///
-/// NOTE: You must not trust the `root` unless you verified it by calling [`verify_account_storage_root`].
-pub fn verify_storage_absence(
-    root: H256,
-    key: U256,
-    proof: impl IntoIterator<Item = impl AsRef<[u8]>>,
-) -> Result<bool, Error> {
-    Ok(get_node(root, key.to_be_bytes(), proof)?.is_none())
 }
 
 /// Computes the execution block root hash.
