+++ lib/evm-storage-verifier/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "evm-storage-verifier"
+version = "0.1.0"
+edition.workspace = true
+license-file.workspace = true
+repository.workspace = true
+
+[dependencies]
+trie-db                     = { version = "0.28", default-features = false }
+thiserror                   = { workspace = true }
+ssz                         = { workspace = true }
+unionlabs                   = { workspace = true, features = ["cosmwasm"] }
+rlp                         = { workspace = true, features = ["derive", "std"] }
+serde-utils = { workspace = true, default-features = false }
+primitive-types             = { workspace = true, features = ["rlp"] }
+memory-db                   = { version = "0.32.0", default-features = false }
+hash-db                     = { version = "0.16.0", default-features = false }
+hash256-std-hasher          = { version = "0.15.2", default-features = false }
+sha3                        = { workspace = true }
+
+[lints]
+workspace = true
