+++ lib/scroll-verifier/Cargo.toml
@@ -12,7 +12,7 @@ workspace = true
 test-include = ["lib/scroll-verifier/tests"]
 
 [dependencies]
-ethereum-verifier         = { workspace = true }
+ evm-storage-verifier         = { workspace = true }
 ethers-core               = { workspace = true }
 hex                       = { workspace = true }
 hex-literal               = { workspace = true }
