+++ light-clients/berachain-light-client/Cargo.toml
@@ -33,7 +33,7 @@ thiserror        = { workspace = true }
 berachain-light-client-types  = { workspace = true, features = ["proto"] }
 cometbft-types                = { workspace = true, features = ["proto", "hash"] }
 ethereum-light-client-types   = { workspace = true, features = ["proto"] }
-ethereum-verifier             = { workspace = true }
+ evm-storage-verifier             = { workspace = true }
 hex-literal                   = { workspace = true }
 ics008-wasm-client            = { workspace = true }
 ics23                         = { workspace = true }
