+++ light-clients/ethereum-light-client/Cargo.toml
@@ -23,7 +23,8 @@ cosmwasm-schema             = { workspace = true }
 cosmwasm-std                = { workspace = true, features = ["abort"] }
 cw-storage-plus             = { workspace = true }
 ethereum-light-client-types = { workspace = true, features = ["proto"] }
-ethereum-verifier           = { workspace = true }
+evm-storage-verifier           = { workspace = true }
+ethereum-sync-protocol = { workspace = true }
 ethers-contract-derive      = { workspace = true }
 ethers-core                 = { workspace = true }
 hex                         = { workspace = true }
