+++ light-clients/scroll-light-client/Cargo.toml
@@ -15,7 +15,7 @@ crate-type = ["cdylib", "rlib"]
 [dependencies]
 cosmwasm-std              = { workspace = true, features = ["abort"] }
 ethereum-light-client     = { workspace = true, features = ["mainnet", "library"] }
-ethereum-verifier         = { workspace = true }
+evm-storage-verifier = { workspace = true }
 ethers-core               = { workspace = true }
 hex                       = { workspace = true }
 ics008-wasm-client        = { workspace = true }
